# editoast/openapi.yaml
openapi: 3.0.2
info:
  title: OSRD Editoast
  description: OSRD Edition service description
  version: 0.1.0
...
paths:
...
  /layers/layer/{layer_slug}/mvt/{view_slug}/:
    get:
      tags:
        - layers
      summary: Mvt View Metadata
      parameters:
        - required: true
          schema:
            title: Layer Slug
            type: string
          name: layer_slug
          in: path
        - required: true
          schema:
            title: View Slug
            type: string
          name: view_slug
          in: path
        - required: true
          schema:
            title: Infra id
            type: integer
          name: infra
          in: query
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewMetadata"
  /layers/tile/{layer_slug}/{view_slug}/{z}/{x}/{y}/:
    get:
      tags:
        - layers
      summary: Mvt tile from the cache or the database, cache data in redis if needed
      parameters:
        - required: true
          schema:
            title: Layer Slug
            type: string
          name: layer_slug
          in: path
        - required: true
          schema:
            title: View Slug
            type: string
          name: view_slug
          in: path
        - required: true
          schema:
            title: Z
            type: integer
          name: z
          in: path
        - required: true
          schema:
            title: X
            type: integer
          name: x
          in: path
        - required: true
          schema:
            title: Y
            type: integer
          name: y
          in: path
        - required: true
          schema:
            title: Infra
            type: integer
          name: infra
          in: query
      responses:
        200:
          description: Successful Response
          content:
            application/x-octet-stream:
              schema:
                type: string
  /infra/:
    get:
      tags:
        - infra
      summary: List all available infra
      responses:
        200:
          description: The infra list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Infra"
    post:
      tags:
        - infra
      summary: Create an infra
      requestBody:
        description: Name of the infra to create
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        201:
          description: The created infra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Infra"
  /infra/{id}/:
    get:
      tags:
        - infra
      summary: Retrieve a specific infra
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: infra id
          required: true
      responses:
        200:
          description: Information about the retrieved infra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Infra"
    delete:
      tags:
        - infra
      summary: Delete an infra and all entities linked to it
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: infra id
          required: true
      responses:
        204:
          description: No content

    post:
      tags:
        - infra
      summary: Update/Create/Delete an object of the infra
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: infra id
          required: true
      requestBody:
        description: Operations to do on the infra
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Operation"
      responses:
        200:
          description: An array containing infos about the operations processed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationResult"

    put:
      tags:
        - infra
      summary: Update an infrastructure name
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: infra id
          required: true
      requestBody:
        description: the name we want to give to the infra
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        200:
          description: The updated infra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Infra"
...
  /infra/{id}/objects/{object_type}/:
    post:
      tags:
        - infra
      summary: "Retrieve a list of specific objects in railjson format"
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Infra id
          required: true
        - in: path
          name: object_type
          schema:
            $ref: "#/components/schemas/ObjectType"
          description: The type of the object
          required: true
      requestBody:
        description: List of object id's
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        200:
          description: No content
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    railjson:
                      type: object
                      description: Object properties in railjson format
                    geographic:
                      type: object
                      description: object's geographic in geojson format
                    schematic:
                      type: object
                      description: object's schematic in geojson format
...
components:
  schemas:
    Infra:
      properties:
        id:
          type: integer
        name:
          type: string
        version:
          type: string
          example: "1"
        generated_version:
          type: string
          nullable: true
          example: "1"
        created:
          type: string
          format: date
        modified:
          type: string
          format: date
        locked:
          type: boolean
...
    ObjectType:
      type: string
      description: Type of the object
      enum:
        - TrackSection
        - Signal
        - SpeedSection
        - Detector
        - TrackSectionLink
        - Switch
        - SwitchType
        - BufferStop
        - Route
        - OperationalPoint
        - Catenary
...
    Operation:
      oneOf:
        - $ref: "#/components/schemas/RailjsonObject"
        - $ref: "#/components/schemas/DeleteOperation"
        - $ref: "#/components/schemas/UpdateOperation"
      discriminator:
        propertyName: operation_type
    OperationResult:
      oneOf:
        - $ref: "#/components/schemas/DeleteOperation"
        - $ref: "#/components/schemas/OperationObject"
      discriminator:
        propertyName: operation_type
    RailjsonObject:
      properties:
        operation_type:
          type: string
          enum: ["CREATE"]
        obj_type:
          $ref: "#/components/schemas/ObjectType"
        railjson:
          $ref: "#/components/schemas/Railjson"
    DeleteOperation:
      properties:
        operation_type:
          type: string
          enum: ["DELETE"]
        obj_type:
          $ref: "#/components/schemas/ObjectType"
        obj_id:
          type: string
          example: bd840b06-84ba-4566-98c1-ccf0196c5f16
    UpdateOperation:
      properties:
        operation_type:
          type: string
          enum: ["UPDATE"]
        obj_type:
          $ref: "#/components/schemas/ObjectType"
        obj_id:
          type: string
          example: bd840b06-84ba-4566-98c1-ccf0196c5f16
        railjson_patch:
          $ref: "#/components/schemas/Patches"
    OperationObject:
      properties:
        operation_type:
          type: string
          enum: ["CREATE", "UPDATE"]
        obj_type:
          $ref: "#/components/schemas/ObjectType"
        railjson:
          $ref: "#/components/schemas/Railjson"
    Patch:
      description: A JSONPatch document as defined by RFC 6902
      required:
        - "op"
        - "path"
      properties:
        op:
          type: string
          description: The operation to be performed
          enum:
            - "add"
            - "remove"
            - "replace"
            - "move"
            - "copy"
            - "test"
        path:
          type: string
          description: A JSON-Pointer
        value:
          type: object
          description: The value to be used within the operations.
        from:
          type: string
          description: A string containing a JSON Pointer value.
    Patches:
      type: array
      description: A list of Patch
      items:
        $ref: "#/components/schemas/Patch"
    Railjson:
      type: object
      description: This field follows railjson format
      example:
        id: bd840b06-84ba-4566-98c1-ccf0196c5f16
        geo:
          type: LineString
          coordinates:
            - - 1.0
              - 41.0
            - - 2.0
              - 42.0
        sch:
          type: LineString
          coordinates:
            - - 1.0
              - 41.0
            - - 2.0
              - 42.0
        curves: []
        length: 1000
        slopes:
          - end: 500
            begin: 250
            gradient: -1
        line_code: 1
        line_name: my line
        track_name: track name
        navigability: BOTH
        track_number: 1
    ViewMetadata:
      properties:
        type:
          type: string
          example: vector
        name:
          type: string
          example: track_sections
        promotedId:
          type: object
          example: { track_sections: "id" }
        scheme:
          type: string
          example: "xyz"
        tiles:
          type: array
          items:
            type: string
            example: http://localhost:7070/tile/track_sections/geo/{z}/{x}/{y}/?infra=1
        attribution:
          type: string
        minzoom:
          type: integer
        maxzoom:
          type: integer
          example: 18
...